+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Improvements
 
 ### Bug Fixes
+* [\#35](https://github.com/Finschia/wasmd/pull/35) stop wrap twice the response of handling non-plus wasm message in plus handler
 
 ### Breaking Changes